requirements.txt
numpy
pandas<2.0.0
scikit-learn
torch
fair-esm
profeatpy
wget
tqdm
fasta